comp 520 assignment: declarations
So we actually had to code it in such a way that there are declarations
as separate from statements, and that declarations can occur at the
beginning of the program followed by the statements, and you can't mix
them. Now that's done, and the unit tests have been adjusted to reflect
this. Thankfully the programs that I wrote did not make use of the
incorrect behaviour previously present. :D