comp 520 minilang: refactor unsafe identifiers into runtime module
This also marks the creation of the Runtime module that will hold the
code that is injected into the generated C code to deal with the
minilang runtime system (RTS). The RTS just provides some functions for
concatenating strings via sprintf in a convenient way and reversing
strings.

If minilang were a more serious language, then the runtime module might
also include garbage collection related stuff.