comp 520 minilang: implement semantic error pretty print
This required removing the overreaching instance
    instance Pretty (f a) => Pretty (Ann x f a)
which would simply pretty print the functorial value ignoring the
annotation entirely in order to write a useful instance
    instance Pretty (Ann SrcSpan SemanticError a)
that actually prints the error location contained in the annotation.

Locations that used the overreaching instance were adjusted to strip the
annotations from the thing they wanted to print. This in particular
required writing the `unannotateProgram` function which simply strips
the annotations from the underlying declarations and statements of the
program.